requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0

# Machine learning
scikit-learn>=1.3.0
scikit-surprise>=1.1.3
scipy>=1.11.0

# Text processing
rapidfuzz>=3.0.0

# API & configuration
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Notebooks
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
ipykernel>=6.25.0

# CLI
click>=8.1.0

# Testing & quality
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.1.0

# Optional: Web dashboard
streamlit>=1.25.0
